moved and verified all the unittests
